# apps/express/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "inventory",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "me",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "kjz-rft=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjAxOTQ0MWJhLWZhMDktNzQ1ZC1hZDYzLWEzMTliMzc2ZmY0MSIsInVwZGF0ZWRfYXQiOm51bGwsImNyZWF0ZWRfYXQiOiIyMDI1LTAxLTA3VDIwOjA2OjAwLjM0MFoiLCJuYW1lIjoiZGVubmlzIiwiZW1haWwiOiJkZW5uaXNraW51dGhpYXdhd2VydUBnbWFpbC5jb20iLCJhdmF0YXJVcmwiOm51bGwsInJvbGUiOiJ1c2VyIiwicmVmcmVzaFRva2VuVmVyc2lvbiI6NiwiaXNFbWFpbFZlcmlmaWVkIjp0cnVlLCJleHAiOjE3MzY3MjI5MjgsImlhdCI6MTczNjI5MDkyOH0.LyVY4wJLyzYUHyFq-EKIIzqrfl4UQaRIUFmdq8pfAy0"
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "no-cache"
          },
          {
            "name": "Postman-Token",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "<calculated when request is sent>"
          },
          {
            "name": "Host",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "<calculated when request is sent>"
          },
          {
            "name": "User-Agent",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "PostmanRuntime/7.39.1"
          },
          {
            "name": "Accept",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "*/*"
          },
          {
            "name": "Accept-Encoding",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "gzip, deflate, br"
          },
          {
            "name": "Connection",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "keep-alive"
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjAxOTQ0MWJhLWZhMDktNzQ1ZC1hZDYzLWEzMTliMzc2ZmY0MSIsInVwZGF0ZWRfYXQiOm51bGwsImNyZWF0ZWRfYXQiOiIyMDI1LTAxLTA3VDIwOjA2OjAwLjM0MFoiLCJuYW1lIjoiZGVubmlzIiwiZW1haWwiOiJkZW5uaXNraW51dGhpYXdhd2VydUBnbWFpbC5jb20iLCJhdmF0YXJVcmwiOm51bGwsInJvbGUiOiJ1c2VyIiwicmVmcmVzaFRva2VuVmVyc2lvbiI6NSwiaXNFbWFpbFZlcmlmaWVkIjp0cnVlLCJleHAiOjE3MzYyOTA5ODgsImlhdCI6MTczNjI5MDkyOH0.xjZOz2DjRJNZP-5fJ1-8xl6HfQFkKUVnZuKvuL2sGlw"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": ["auth"],
        "summary": "verify-email",
        "requestBody": {
          "content": {}
        },
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "denniskinuthiaw@gmail.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}